+++ src/main/java/com/kinsolutions/users/service/UserService.java
@@ -4,6 +4,6 @@ import com.kinsolutions.model.Users;
 
 public interface UserService {
 	
-	public Users getUserByUserId(long userId);
+	public Users getUserByUserId(Integer userId);
 
 }
